Puse las cosas en la base de datos pero hay un error que andres quiere solucionar

+++ src/main/java/com/vetcare/proyecto/Exepciones/NotFoundException.java
@@ -1,16 +1,16 @@
 package com.vetcare.proyecto.Exepciones;
 
 public class NotFoundException extends RuntimeException{
-    private Integer id;
+    private Long id;
 
-    public NotFoundException(Integer id){
+    public NotFoundException(Long id){
         this.id = id;
     }
 
-    public Integer getId(){
+    public Long getId(){
         return id;
     }
-    public void setId(Integer id){
+    public void setId(Long id){
         this.id = id;
     }
