+++ src/main/java/com/vetcare/proyecto/service/MascotaServicio.java
@@ -6,15 +6,15 @@ import java.util.Collection;
 import com.vetcare.proyecto.entities.Mascota;
 
 public interface MascotaServicio {
-    public Mascota GetById(int id);
+    public Mascota GetById(Long id);
 
     public Collection<Mascota> GetAll();
 
     public void addMascota(Mascota mascota);
 
-    public void removeMascota(int id);
+    public void removeMascota(Long id);
 
     public void updateMascota(Mascota mascota);
 
-    public ArrayList<Mascota> GetVariousById(ArrayList<Integer> ids);
+    public ArrayList<Mascota> GetVariousById(ArrayList<Long> ids);
 }
